Fix bugs and add improvements: Type mismatch fix, pagination, database indexes - fixes EnrollmentResponse.progress type (Double→Integer), adds full pagination support for notifications (page/size params, Page<T> return type), adds database @Index annotations to Certificate/CourseReview/Category/Notification for performance optimization (BUILD SUCCESS)

+++ Vanca/src/main/java/com/example/demo/dto/response/EnrollmentResponse.java
@@ -18,5 +18,5 @@ public class EnrollmentResponse {
 	private Long studentId;
 	private String studentName;
 	private LocalDateTime enrolledAt;
-	private Double progress; // Progress percentage (0-100)
+	private Integer progress; // Progress percentage (0-100)
 }
